feat(Colorscreens): Add color hightlights support
If u have a color screen like the kobo libra colour you could configure every color to have an specific hashtag and then search for that hashtag in your obsidian to get all the red hightlights. This is very useful if you use some sort of color code when highlight.

Signed-off-by: nothingbutlucas <69118979+nothingbutlucas@users.noreply.github.com>